+++ Assets/CodeBase/Runtime/Interfaces/ITool.cs
@@ -2,6 +2,6 @@ using System;
 public interface ITool
 {
     //public int Index { get; } 
-    public int BaseDamage { get; }
-    public int CalculateDamage(Type hitObjectType);
+    public float BaseDamage { get; }
+    public float CalculateDamage(Type hitObjectType);
 }
